news-aggregator-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.119.0"}
psycopg = {extras = ["binary"], version = "^3.2.10"}
feedparser = "^6.0.12"
beautifulsoup4 = "^4.14.2"
requests = "^2.32.5"
python-dateutil = "^2.9.0.post0"
lxml = "^6.0.2"
selenium = "^4.36.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
